Implementing the editor by searching: laravel-ckeditor. And then install it by: composer require unisharp/laravel-ckeditor.
And then add the provider:
Unisharp\Ckeditor\ServiceProvider::class,
to the app.php in the config folder.

After that we need to publish to the resources by:
php artisan vendor:publish --tag=ckeditor.

Then we should add Usage in to the app.blade.php in the layout folder.

And we need to add the 'id'=> 'article-ckeditor' to the textarea form array in the create.blade.php file in the posts folder.