We need to edit and delete posts. So we are going to add anchor tag to take us to the edit file in the show.blade.php file.
After that we should go to the PostsController.php file to add  find id and return the edit view to the edit function.

After creating the edit.blade.php file in the posts folder we copy paste all from the create.blade.php file to edit.blade.php file but we should change the heading and the action to update and also we need to include the id via array.

We need to add PUT _method via hidden form for at Laravel can understand at this is update request.

We need to add the value instead at being blank to edit them and we need to copy paste all from the store request to the update request but we should change the Post model from new to find($id), plus the message to updated.