Adding relationships between the blog posts  and users for at being able to display just to the current user his belong posts in the dashboard. So in the Post model class we add user function which:  return $this->belongsTo('App/User');
And in the other side we create posts function which:
return $this->hasMany('App/posts'); //it’s one to many relationship//

Now that’s will make us able to use certain queries and certain elements that normally would not be available to us.

In the dashboard we want to fetch the post for the specific user.
And in the dashboard.blade.php file in the view we create a table to show the posts belong.
And don't forget to bring the user model in to the DashboardController by:
use App\User;